+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,52 +1,10 @@
 package com.techcourse.service;
 
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import org.springframework.dao.DataAccessException;
 
-public class UserService {
+public interface UserService {
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public User findById(final Connection connection, final long id) {
-        return userDao.findById(connection, id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) throws SQLException {
-        final var connection = DataSourceConfig.getInstance().getConnection();
-        try {
-            connection.setAutoCommit(false);
-            final var user = findById(connection, id);
-            user.changePassword(newPassword);
-            userDao.update(connection, user);
-            userHistoryDao.log(connection, new UserHistory(user, createBy));
-
-            connection.commit();
-        } catch (final Exception e) {
-            connection.rollback();
-            throw new DataAccessException(e);
-        } finally {
-            connection.close();
-        }
-    }
-
+    User findById(final long id);
+    void insert(final User user);
+    void changePassword(final long id, final String newPassword, final String createBy);
 }
